pyproject.toml
[project]
name = "knuckledragger"
version = "0.1.1"
authors = [{ name = "Philip Zucker", email = "philzook58@gmail.com" }]
description = "Interactive Theorem Prover"
readme = "README.md"
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
requires-python = ">= 3.12"
dependencies = [
    "z3-solver >= 4.8",
    "egglog >= 9.0.0",
    "python-flint",
    "sympy",
    "lark",
    "hypothesis>=6.124.1",
    "cffi>=1.17.1",
    "graphviz>=0.20.3",
    "click>=8.1.8",
]
[project.urls]
homepage = "https://github.com/philzook58/knuckledragger"


# For raw pip? 
[project.optional-dependencies]
dev = ["pytest>=6.0", "nbclient", "jupyter"]
yosys = ["yowasp-yosys>=0.50.0.0.post858"]
pypcode = ["cle>=9.2.148", "pypcode>=3.0.3"]
rust = ["maturin>=1.8.6"]

[tool.setuptools]
include-package-data = true
#package-dir = { "" = "kdrag" }
#packages = ["kdrag"] # automatic discovery is better
[tool.setuptools.package-data]
knuckledragger = ["*.sh"]

[tool.ruff]
exclude = ["tests"]

[tool.ruff.lint]
ignore = ["E741"]

[tool.ruff.lint.per-file-ignores]
"src/kdrag/smt.py" = ["F405", "F403"]
"src/kdrag/all.py" = ["F401"]
"tutorial.ipynb" = ["F811"]
"examples/**/*.py" = ["F401", "F405", "F403", "E402"]
"examples/short_talk.ipynb" = ["F401"]

[tool.pyright]
include = ["src"]
exclude = [
    "src/kdrag/solvers/eprover/*",
    "src/kdrag/solvers/gappa*",
    "src/kdrag/solvers/Prover9/*",
]

[tool.ty.src]
include = ["src"]
exclude = [
    "src/kdrag/solvers/eprover/*",
    "src/kdrag/solvers/gappa*",
    "src/kdrag/solvers/Prover9/*",
]

[tool.pytest.ini_options]
norecursedirs = [
    "rust/target",
    "solvers/eprover",
    "solvers/gappa*",
    "solvers/Prover9",
    "examples/converted_notebooks",
]
addopts = "--doctest-modules"
doctest_optionflags = "ELLIPSIS NORMALIZE_WHITESPACE IGNORE_EXCEPTION_DETAIL"

markers = ["slow: Marks tests as long-running."]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "nbclient>=0.10.2",
    "pyright>=1.1.392.post0",
    "pytest>=8.3.4",
    "ruff>=0.11.13",
]
rust = ["maturin[patchelf]>=1.8.6"]
